+++ mycodo/config.py
@@ -92,7 +92,12 @@ CONDITIONAL_CONDITION_INFO = {
 
 # Conditional controllers
 CONDITIONAL_CONDITIONS = [
-    ('measurement', TRANSLATIONS['measurement']['title']),
+    ('measurement', "{} ({})".format(
+        TRANSLATIONS['measurement']['title'],
+        TRANSLATIONS['single']['title'])),
+    ('measurement_dict', "{} ({})".format(
+        TRANSLATIONS['measurement']['title'],
+        TRANSLATIONS['multiple']['title'])),
     ('gpio_state', lazy_gettext('GPIO State')),
     ('output_state', lazy_gettext('Output State'))
 ]
@@ -697,6 +702,7 @@ PID_INFO = {
         }
     }
 }
+
 
 FUNCTION_INFO = {
     'function_spacer': {