+++ mycodo/controller_conditional.py
@@ -40,7 +40,6 @@ from mycodo.databases.models import Misc
 from mycodo.databases.models import SMTP
 from mycodo.mycodo_client import DaemonControl
 from mycodo.utils.database import db_retrieve_table_daemon
-from mycodo.utils.system_pi import is_int
 
 MYCODO_DB_PATH = 'sqlite:///' + SQL_DATABASE_MYCODO
 
@@ -90,6 +89,7 @@ class ConditionalController(threading.Thread):
         self.period = None
         self.start_offset = None
         self.refractory_period = None
+        self.log_level_debug = None
         self.conditional_statement = None
         self.timer_refractory_period = None
         self.smtp_wait_timer = None
@@ -181,9 +181,15 @@ class ConditionalController(threading.Thread):
         self.period = cond.period
         self.start_offset = cond.start_offset
         self.refractory_period = cond.refractory_period
+        self.log_level_debug = cond.log_level_debug
         self.timer_refractory_period = 0
         self.smtp_wait_timer = now + 3600
         self.timer_period = now + self.start_offset
+
+        if self.log_level_debug:
+            self.logger.setLevel(logging.DEBUG)
+        else:
+            self.logger.setLevel(logging.INFO)
 
     def check_conditionals(self):
         """
@@ -244,6 +250,15 @@ message='''{message}'''
 
 def measure(condition_id):
     return control.get_condition_measurement(condition_id)
+
+def measure_dict(condition_id):
+    string_sets = control.get_condition_measurement_dict(condition_id)
+    if string_sets:
+        list_ts_values = []
+        for each_set in string_sets.split(';'):
+            ts_value = each_set.split(',')
+            list_ts_values.append({{'time': ts_value[0], 'value': float(ts_value[1])}})
+        return list_ts_values
 
 def run_all_actions(message=message):
     control.trigger_all_actions('{function_id}', message=message)