+++ mycodo/inputs/ttn_data_storage.py
@@ -47,12 +47,14 @@ INPUT_INFORMATION = {
     'input_name': 'TTN Integration: Data Storage',
     'measurements_name': 'Variable measurements',
     'measurements_dict': measurements_dict,
+    'measurements_use_same_timestamp': False,
 
     'options_enabled': [
         'custom_options',
         'measurements_select',
         'period',
-        'pre_output'
+        'pre_output',
+        'log_level_debug'
     ],
     'options_disabled': ['interface'],
 
@@ -118,6 +120,11 @@ class InputModule(AbstractInput):
                     elif option == 'device_id':
                         self.device_id = value
 
+        if input_dev.log_level_debug:
+            self.logger.setLevel(logging.DEBUG)
+        else:
+            self.logger.setLevel(logging.INFO)
+
     def get_new_data(self, past_seconds):
         # Basic implementation. Future development may use more complex library to access API
         endpoint = "https://{app}.data.thethingsnetwork.org/api/v2/query/{dev}?last={time}".format(
@@ -173,7 +180,7 @@ class InputModule(AbstractInput):
                     measurements[each_meas.channel]['measurement'] = each_meas.measurement
                     measurements[each_meas.channel]['unit'] = each_meas.unit
                     measurements[each_meas.channel]['value'] = each_resp[each_meas.name]
-                    measurements[each_meas.channel]['timestamp'] = datetime_utc
+                    measurements[each_meas.channel]['timestamp_utc'] = datetime_utc
 
                     # Convert value/unit is conversion_id present and valid
                     if each_meas.conversion_id:
@@ -190,9 +197,10 @@ class InputModule(AbstractInput):
                             measurements[each_meas.channel]['measurement'] = meas[each_meas.channel]['measurement']
                             measurements[each_meas.channel]['unit'] = meas[each_meas.channel]['unit']
                             measurements[each_meas.channel]['value'] = meas[each_meas.channel]['value']
-                            measurements[each_meas.channel]['timestamp'] = datetime_utc
 
-            add_measurements_influxdb(self.unique_id, measurements)
+            add_measurements_influxdb(
+                self.unique_id, measurements,
+                use_same_timestamp=INPUT_INFORMATION['measurements_use_same_timestamp'])
 
         # set datetime to latest timestamp
         if self.running: